Fixed issue with USB-C pad spacing on 2oz copper
The minimum spacing between tracks on 2oz copper is 0.16mm, and I had explicitly ignored the USB-C connector pin spacing 0.12mm in a custom DRC rule, which turned out to be a problem. I removed that rule and modified the USB-C connector pads to be slightly narrower to give more gap.